fix dismiss func
